Use rerun gem for live reloading in development mode
Previously we used Sinatra::Reloader to live reload the app on
changes. This required an additional action from the user to add a line of
code in each file that needs to be reloaded.

Rerun restarts the complete webrick server, requiring more time to
restart than Sinatra::Reloader, but it doesn't require an additional
action from the user.